Autooptimize & constrain no. of PhantomJS instances
Add an option to automatically calculate the best number of PhantomJS
instances (if `--phantomjs-instances auto` is passed to the script),
and to fix incorrect / too big input values.

The CLI parameter is required, because the default is 0 (don't use
Phantom).

The number of instances will be constrained to available CPU threads and
free memory. Exceeding the resources results mostly in decreased
performance and/or crashes.

Before that, there was a way to set the number of instances by invoking
e.g.

`npm config set ariatemplates:phantomjsInstances 8`

but this worked only when running attester via `npm run-script` (the
default value in `package.json` had to be 2 due to limitation of Travis
CI).

Close #30.