feat #32 default port 7777; check if port available
Attester used to let Node decide about the port for the server if it was
not passed as a parameter. This however is not very handy to have random
port on each run.

This commits makes 7777 be the default port (unless some value is
explicitly passed, or 0 for the old, random behavior).

Also, if the passed port number (or default one) is taken, the next
available port number is used instead and a warning is printed.

Close #32.