Better logging on the slave
Redesign the slave interface to have a clear understanding on what's happening on the slave.
This is obtained enabling console.log and writing logs in a div when the URL parameter `log` is set to `true`.

If also adds additional elements in the page that can be used to query the status of the slave with tools like selenium.
An invisilbe element with id `status_info` contains another element with id

`_info_[status]`

Possible values for `status` are:
* `loading` when the slave is trying to connect
* `connected` or `disconnected` as soon as socket.io establishes or looses a connection
* `executing` while running a test
* `idle` after a test has been executed