Fix missing console logs from TestCampaign constructor
After refactoring attester as a module, logger is attached too late for
the logs from campaign constructor to be displayed in the console (because
"attester.start" is raised after the initial campaigns creation).

This adds a new event "attester.config.available" which is raised at the
very beginning, before campaign creation, so we can set up the logging
early.

Close #51.