More flexible `extraScripts`
Now `extraScripts` can either be an array (legacy mode; inserted into test page
after the bootstrap for AT tests) or an object with `before` and `after` nodes,
specifying insertion order with respect to the bootstrap.

This commits reverts the default behavior of legacy mode to add files
*before* the bootstrap (as was changed in 5c6e86606fa306f22652bc1a3f711191defdbe5a)
to the old behavior of 1.2.1 which is to add them *after* the bootstrap
(which was used mainly to include the atskin after the framework).

Internally, in JSON files describing the test page templates, instead
of just `head`, also `head-start` and `head-end` are available now.

Close #55. Close #58.