Maintenance
This commit does a number of little fixes:

- fix a problem with undefined PhantomJS id
- do not reboot PhantomJS when received SIGTERM
- add `grunt-cli` file and update package.json so that we can
  pass parameters to node executable if needed
- remove `shortCut` since it's only used in one place and adds
  unnecessary complexity
- added some debug logging
- removed unused import

This commit also changes order of loading attester modules, so that
when they're unloaded in __reset__, it's done in an order that gives
more chance for the websocket connection with PhantomJS to be closed
before attester kills the Phantom child processes.

However there's still a timing issue to be worked around properly
(since things are asynchronous both on server and client), hence in
Node 0.10 it's still likely to see this error message when disposing
attester:

  [attester.TestServer.socketio] socket error Error: read ECONNRESET
      at errnoException (net.js:904:11)
      at TCP.onread (net.js:558:19)

Note that this timing issue likely doesn't occur when attester is started
with debug log level, due to this call in the code:

  attester.logger.logDebug("__reset__ " + module);

Then the server is a bit more busy and the client has more time
to disconnect properly.