fix #16 task restart mechanism when a browser is disconnected
Before this commit, when a browser was disconnected while executing a
task, the task was considered as failed.
This commit automatically queues again those interrupted tasks so that
they are executed again (with a maximum number of restarts, so that
tasks which make a browser crash do not lead to a never ending campaign).

Close #100