fix issue with events order in edge cases (campaignFinished before serverAttached)
This commit makes sure the campaignFinished event never happens before
calling the callback of campaign.init. This could previously happen if,
for example, all the classpaths in the campaign were invalid.

Having another event after the campaignFinished event (such as the
serverAttached event) lead to an exception in json-log-report.js (because
_stream is set to null when receiving the campaignFinished event). Now,
any event happening after campaignFinished is ignored in json-log-report.