Fixing random issues in phantom.spec.js
The test asserts the number of messages in the output of the phantomjs
control script, but each call to the listener of the data event (of the
output stream) created a single new message. This was not reliable as
the data event listenr can sometimes be called once for several
concatenated messages.