Using the default nodejs hostname instead of 0.0.0.0 when not specified
Now instead of using 0.0.0.0 as the default hostname value, this commit
lets nodejs decide which is the best hostname to use in case no --host
parameter is specified on the command line. Especially, when listening
on 0.0.0.0 and using the name of the machine in the URL, requests can be
very slow if the machine also has an IPv6 address, because the browser
will usually first try to connect to the IPv6 address and then fallback
to the IPv4 address after some time (which can be quite long).
On nodejs >= 0.12, a server listening on the "::" address (which is the
default if IPv6 is detected) is also reachable by the IPv4 addresses, so
detectHostname.js had to be updated to also use the IPv4 addresses to
find the correct host name.