Adding taskExecutionId to check the link between a task and its updates
In some cases, a task can impact the following one by reporting updates
after the taskFinished event.
This commit adds an id which is different for each task execution and
checks it with every test update so that it is sure that an update was
correcly raised by the right task.
On the browser, a new instance of the attester API is created for each
task, so that when a task is calling the API, attester knows from which
task execution the call is coming and can avoid reporting updates for
already finished tasks.