Reporting console logs to attester.
With this commit, when using one of the console methods (log, info, warn
or error) from a test in any browser, the log is now reported to the
attester server and displayed on its console for the corresponding slave.
If the console method is called from the currently running task, the
corresponding log events are also included in the campaign logs (to
be displayed in attester-results-ui).