Switching to sockjs instead of socket.io (non-backward compatible change)
This commit makes attester use sockjs instead of socket.io, which is a
lower-level library that seems to work better, especially with node.js 8.x.

This changes the protocol, so the new version of attester will not work
with old versions of attester-launcher and attester-results-ui and the new
versions of attester-launcher and attester-results-ui will not work with
old versions of attester.

This change makes attester compatible with node.js 8.2.

It also removes support for older versions of node.js. That is especially
caused by the new version of selenium-webdriver used by attester-launcher
which is not compatible with old versions of node.js.